requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1

# AI and Code Analysis
ollama==0.0.50
tree-sitter==0.20.3
tree-sitter-languages==1.10.2
anthropic==0.7.0

# Screen capture and OCR
pillow==10.1.0
pyscreenshot==3.1
mss==9.0.1
pytesseract==0.3.10

# Code execution and security
subprocess-tee==0.4.1
watchdog==3.0.0
pygments==2.17.2

# File operations
pathlib2==2.3.7.post1

# Logging and monitoring
python-json-logger==2.0.7

# Web and API
requests==2.31.0
httpx==0.25.2

# Development
pytest==7.4.3
black==23.12.0
flake8==6.1.0
mypy==1.7.1
